Change influence of tile dim size in offset_mapper
In case of fixed tile sizes, while the sizes would be pre-multiplied in
the TIF, when calculating contiguous 2D offset they are not, i.e.
something like lsc_offset({},[],[],1) needs to be multiplied with not
just M/N but also tile.dima.size (or tile.dimb.size depending on
vecdir). Another way to see this is that the base offset should be
lsc_offset({},[],[],tile.dimX.size) instead of lsc_offset({},[],[],1)
and then this base should be multiplied with the dimension